@@ -1,64 +1,55 @@
 {
   "name": "annotation-app",
-  "version": "0.10.0",
-  "description": "Angular-CLI starter for a Clarity project",
-  "angular-cli": {},
+  "version": "0.0.0",
   "scripts": {
-    "start": "ng serve --proxy-config proxy.conf.json",
-    "lint": "tslint \"src/**/*.ts\" -e \"src/polyfills.ts\" --force",
-    "prettier": "npx prettier --write .",
-    "build": "ng build"
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "bundle": "ng run annotation-app:mfe && node build-script.js",
+    "bundle_dev": "ng run annotation-app:mfe:dev && node build-script.js",
+    "bundle_stg": "ng run annotation-app:mfe:stg && node build-script.js",
+    "bundle_prod": "ng run annotation-app:mfe:prod && node build-script.js"
   },
   "private": true,
   "dependencies": {
-    "@angular-slider/ngx-slider": "2.0.3",
-    "@angular/animations": "8.2.13",
-    "@angular/common": "8.2.13",
-    "@angular/compiler": "8.2.13",
-    "@angular/core": "8.2.13",
-    "@angular/forms": "8.2.13",
-    "@angular/platform-browser": "8.2.13",
-    "@angular/platform-browser-dynamic": "8.2.13",
-    "@angular/router": "8.2.13",
-    "@auth0/angular-jwt": "3.0.1",
-    "@clr/angular": "2.2.0",
-    "@clr/icons": "2.2.0",
-    "@clr/ui": "2.2.0",
-    "@ng-select/ng-select": "3.7.3",
-    "@webcomponents/custom-elements": "^1.0.0",
-    "aws-sdk": "2.814.0",
-    "buffer": "5.5.0",
-    "core-js": "2.4.1",
+    "@angular-devkit/build-angular": "^14.2.9",
+    "@angular-slider/ngx-slider": "^2.0.3",
+    "@angular/animations": "^14.2.0",
+    "@angular/common": "^14.2.0",
+    "@angular/compiler": "^14.2.0",
+    "@angular/core": "^14.2.0",
+    "@angular/elements": "^14.2.11",
+    "@angular/forms": "^14.2.0",
+    "@angular/platform-browser": "^14.2.0",
+    "@angular/platform-browser-dynamic": "^14.2.0",
+    "@angular/router": "^14.2.0",
+    "@auth0/angular-jwt": "^3.0.1",
+    "@cds/angular": "6.2.2",
+    "@cds/core": "6.2.2",
+    "@clr/angular": "13.10.1",
+    "@clr/ui": "13.10.1",
+    "@ng-select/ng-select": "^9.1.0",
+    "@types/lodash": "^4.14.189",
+    "aws-sdk": "^2.1261.0",
+    "buffer": "^5.5.0",
+    "concat": "^1.0.3",
+    "d3-v6-tip": "^1.0.9",
+    "fs-extra": "^10.1.0",
     "js-untar": "^2.0.0",
-    "jszip": "3.7.1",
-    "label-studio": "0.8.0",
-    "lodash": "4.17.21",
-    "marked": "2.1.3",
-    "mutationobserver-shim": "0.3.2",
-    "ng2-charts": "1.6.0",
-    "ngx-papaparse": "4.0.4",
+    "jszip": "^3.7.1",
+    "lodash": "^4.17.21",
+    "marked": "^2.1.3",
+    "ngx-build-plus": "^14.0.0",
+    "ngx-papaparse": "^5.1.0",
     "pako": "^2.0.3",
-    "rxjs": "6.4.0",
-    "tslib": "1.10.0",
-    "web-animations-js": "2.3.2",
-    "zone.js": "0.9.1"
+    "rxjs": "~7.5.0",
+    "tslib": "^2.3.0",
+    "windowed-observable": "^1.0.0",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.29",
-    "@angular/cli": "8.3.17",
-    "@angular/compiler-cli": "8.2.13",
-    "@types/core-js": "0.9.42",
-    "@types/node": "8.0.24",
-    "bootstrap": "4.0.0-alpha.5",
-    "codelyzer": "6.0.2",
-    "enhanced-resolve": "3.4.1",
-    "forever": "3.0.4",
-    "jquery": "3.6.0",
-    "prettier": "2.3.1",
-    "ts-node": "4.1.0",
-    "tslint": "6.1.3",
-    "tslint-angular": "3.0.3",
-    "tslint-config-prettier": "1.18.0",
-    "typescript": "3.5.3"
+    "@angular/cli": "~14.2.5",
+    "@angular/compiler-cli": "^14.2.0",
+    "typescript": "~4.7.2"
   }
 }
