@@ -1,5 +1,5 @@
 /*
-Copyright 2019-2021 VMware, Inc.
+Copyright 2019-2023 VMware, Inc.
 SPDX-License-Identifier: Apache-2.0
 */
 
@@ -12,13 +12,17 @@ import { EnvironmentsService } from '../app/services/environments.service';
 export function jwtOptionsFactory(authService: UserAuthService, env: EnvironmentsService) {
   return {
     tokenGetter: () => {
+      // console.log('tokenGetter: ()---', authService.loggedUser());
       if (authService.loggedUser()) {
         return authService.loggedUser().token ? authService.loggedUser().token.access_token : null;
       }
+      return null;
     },
     whitelistedDomains: [
-      `${env.config.annotationService}`.replace(/(http|https):\/\//, ''),
+      `${env.config.annotationService}`.replace(/(http|https):\/\//, '').split('/')[0],
       `${env.config.authUrl}`.replace(/(http|https):\/\//, ''),
+      `${env.config.inUrl}`.substring(0, `${env.config.inUrl}`.length - 8).replace(/(http|https):\/\//, ''),
+      `${env.config.hubService}`.substring(0, `${env.config.hubService}`.length - 8).replace(/(http|https):\/\//, ''),
     ],
   };
 }
