@@ -1,5 +1,5 @@
 /*
-Copyright 2019-2021 VMware, Inc.
+Copyright 2019-2023 VMware, Inc.
 SPDX-License-Identifier: Apache-2.0
 */
 
@@ -31,38 +31,29 @@ export interface AuthResponseToken {
 }
 
 export interface AuthUser {
-  email?: string;
-  provider?: string;
-  sub?: string;
-  username?: string;
-  exp?: number;
-  iat?: number;
-  iss?: string;
-  name?: string;
-  ob_sso_cookie?: string;
   token?: AuthResponseToken;
-
-  id?: string;
-  btoa?: string;
-  fullName?: string;
-  srs?: number[];
-  srCount?: number;
-  optOutProducts?: string[];
-  points?: number;
-  percentage?: number;
-  role?: any;
+  user?: {
+    role?: any;
+    email?: string;
+    provider?: string;
+    sub?: string;
+    username?: string;
+    exp?: number;
+    iat?: number;
+    iss?: string;
+    name?: string;
+  };
 }
 
 export class AuthUtil {
   public static isValidUser(user: AuthUser): boolean {
-    if (user && user.email && user.token && user.token.access_token && user.token.refresh_token) {
+    if (user && user.user.email && user.token && user.token.access_token && user.token.refresh_token) {
       return true;
     }
     return false;
   }
-
-  public static isValidBasicUser(user: AuthUser): boolean {
-    if (user && user.email && user.token && user.token.access_token && user.token.expires_time) {
+  public static isValidBasicUser(user): boolean {
+    if (user && user.user.email && user.token && user.token.access_token && user.token.expires_time) {
       return true;
     }
     return false;
