@@ -1,9 +1,10 @@
 /*
-Copyright 2019-2021 VMware, Inc.
+Copyright 2019-2023 VMware, Inc.
 SPDX-License-Identifier: Apache-2.0
 */
 
 export interface Env {
+  embedded?: boolean;
   production?: boolean;
   annotationService?: string;
   redirectUrl?: string;
@@ -16,11 +17,16 @@ export interface Env {
   tokenUrl?: string;
   logoutUrl?: string;
   CLIENT_ID?: string;
-  feedbackUrl?: string;
-  videoSrc?: string;
-  googleTrackId?: string;
+  lumosUrl?: string;
+  googleTrackId?: any;
   enableAWSS3?: boolean;
   contactEmail?: string;
   contactSlack?: string;
   fileSize?: number;
+  enableSlack?: boolean;
+  slackAppName?: string;
+  sessionKey?: string;
+  inUrl?: string;
+  videoSrc?: string;
+  hubService?: string;
 }
