@@ -1,5 +1,5 @@
 /*
-Copyright 2019-2021 VMware, Inc.
+Copyright 2019-2023 VMware, Inc.
 SPDX-License-Identifier: Apache-2.0
 */
 
@@ -14,3 +14,5 @@ export interface User {
   points?: number;
   percentage?: number;
 }
+
+export const LabelStudio = undefined;
