@@ -1,36 +1,33 @@
 /*
-Copyright 2019-2021 VMware, Inc.
+Copyright 2019-2023 VMware, Inc.
 SPDX-License-Identifier: Apache-2.0
 */
 
 import { Injectable } from '@angular/core';
 import * as _ from 'lodash';
-import { ToolService } from 'app/services/common/tool.service';
-import { AvaService } from '../ava.service';
+import { ToolService } from 'src/app/services/common/tool.service';
+import { ApiService } from '../api.service';
+import { InternalApiService } from '../internal-api.service';
 
 @Injectable()
 export class EmailService {
-  constructor(private toolService: ToolService, private avaService: AvaService) {}
+  constructor(
+    private toolService: ToolService,
+    private apiService: ApiService,
+    private internalApiService: InternalApiService,
+  ) {}
 
   public sendEmailToAnnotator(param) {
-    this.avaService.sendEmailToAnnotator(param).subscribe(
-      (res) => {
-        console.log('sendEmailToAnnotator:::', res);
-      },
-      (error: any) => {
-        console.log(error);
-      },
+    this.internalApiService.sendEmailToAnnotator(param).subscribe(
+      (res) => {},
+      (error: any) => {},
     );
   }
 
   public sendEmailToOwner(param) {
-    this.avaService.sendEmailToOwner(param).subscribe(
-      (res) => {
-        console.log('sendEmailToOwner:::', res);
-      },
-      (error: any) => {
-        console.log(error);
-      },
+    this.internalApiService.sendEmailToOwner(param).subscribe(
+      (res) => {},
+      (error: any) => {},
     );
   }
 
