@@ -1,5 +1,5 @@
 /*
-Copyright 2019-2021 VMware, Inc.
+Copyright 2019-2023 VMware, Inc.
 SPDX-License-Identifier: Apache-2.0
 */
 
@@ -9,21 +9,17 @@ export class DatasetValidator {
   private static readonly REQUIRED_FIELD: string = 'This field is required';
   private static readonly REQUIRED_FIELD_LABEL: string = 'This field is required at least 2 labels';
   private static readonly REQUIRED_FIELD_POP_LABEL: string = 'This field is required at least 2 secondary labels';
-  private static readonly REQUIRED_FIELD_ENTITY: string =
-    'This field is required at least 1 entity';
+  private static readonly REQUIRED_FIELD_ENTITY: string = 'This field is required at least 1 entity';
   private static readonly FILE_FORMAT_NOT_SUPPORT: string = 'Selected file format is not supported';
-  private static readonly FILE_SIZE_EXCEED_LIMIT: string =
-    'Selected file size exceeds the limit 500MB';
-  private static readonly CREATE_FILE_SIZE_EXCEED_LIMIT: string =
-    'Selected file size exceeds the limit 100MB. Please use the My Datasets tab for larger datasets. Once completed, data will be available in this menu';
-  private static readonly IMAGE_FILE_SIZE_EXCEED_LIMIT: string =
-    'Selected file size exceeds the limit 1MB';
+  private static readonly FILE_SIZE_EXCEED_LIMIT: string = 'Selected file size exceeds the limit 500MB';
+  private static readonly IMAGE_FILE_SIZE_EXCEED_LIMIT: string = 'Selected file size exceeds the limit 1MB';
   private static readonly FILE_DUPLICATED: string = 'Selected file has already exist in database';
-
+  private static readonly FILE_EXCEEDS_100MB: string =
+    'File exceeds 100MB. Please use the My Datasets tab for larger dataset upload. Once completed, data will be available in this menu.';
   static modelName(): ValidatorFn {
     return (control: AbstractControl): { [key: string]: any } => {
       if (!control.parent) {
-        return null;
+        return null!;
       }
 
       if (DatasetValidator.isEmpty(control.value)) {
@@ -35,19 +31,18 @@ export class DatasetValidator {
       if (!control.value.match(argRegEx)) {
         return {
           msg: {
-            value:
-              'Wrong format! Model name only allow letters, digits, dots, underscores and hyphen',
+            value: 'Wrong format! Model name only allow letters, digits, dots, underscores and hyphen',
           },
         };
       }
-      return null;
+      return null!;
     };
   }
 
   static datasetName(): ValidatorFn {
     return (control: AbstractControl): { [key: string]: any } => {
       if (!control.parent) {
-        return null;
+        return null!;
       }
 
       if (DatasetValidator.isEmpty(control.value)) {
@@ -59,19 +54,18 @@ export class DatasetValidator {
       if (!control.value.match(argRegEx)) {
         return {
           msg: {
-            value:
-              'Wrong format! Dataset name only allow digits, letters, dots, underscore, and hyphen',
+            value: 'Wrong format! Dataset name only allow digits, letters, dots, underscore, and hyphen',
           },
         };
       }
-      return null;
+      return null!;
     };
   }
 
   static validEmail(): ValidatorFn {
     return (control: AbstractControl): { [key: string]: any } => {
       if (!control.parent) {
-        return null;
+        return null!;
       }
 
       if (DatasetValidator.isEmpty(control.value)) {
@@ -81,16 +75,18 @@ export class DatasetValidator {
       const pattern = `@vmware.com\\s*$`;
       const argRegEx = new RegExp(pattern, 'g');
       if (!control.value.match(argRegEx)) {
-        return { msg: { value: 'Wrong format! Email only accept vmware emailbox' } };
+        return {
+          msg: { value: 'Wrong format! Email only accept vmware emailbox' },
+        };
       }
-      return null;
+      return null!;
     };
   }
 
   static validNormalEmail(): ValidatorFn {
     return (control: AbstractControl): { [key: string]: any } => {
       if (!control.parent) {
-        return null;
+        return null!;
       }
 
       if (DatasetValidator.isEmpty(control.value)) {
@@ -99,7 +95,7 @@ export class DatasetValidator {
 
       const pattern = /^[^\s@]+@[^\s@]+\.[^\s@]{2,}$/;
       if (pattern.test(control.value)) {
-        return null;
+        return null!;
       } else {
         return { msg: { value: 'Wrong format! Only accept email address' } };
       }
@@ -109,30 +105,20 @@ export class DatasetValidator {
   static validPassword(): ValidatorFn {
     return (control: AbstractControl): { [key: string]: any } => {
       if (!control.parent) {
-        return null;
+        return null!;
       }
 
       if (DatasetValidator.isEmpty(control.value)) {
         return { msg: { value: DatasetValidator.REQUIRED_FIELD } };
       }
-
-      // const argRegEx = new RegExp('(?=.*[0-9])(?=.*[A-Z])(?=.*[a-z])(?=.*[^a-zA-Z0-9]).{8,20}');
-      // if (!control.value.match(argRegEx)) {
-      //   return {
-      //     msg: {
-      //       value:
-      //         'Use at least 8 and at most 20 characters, including at least one of each of the following: special character (@%+/’!#$^?:,()[]~-_.), lowercase, uppercase and number.',
-      //     },
-      //   };
-      // }
-      return null;
+      return null!;
     };
   }
 
   static validRow(): ValidatorFn {
     return (control: AbstractControl): { [key: string]: any } => {
       if (!control.parent) {
-        return null;
+        return null!;
       }
       if (DatasetValidator.isEmpty(control.value)) {
         // return { 'msg': { value: DatasetValidator.REQUIRED_FIELD } };
@@ -142,16 +128,16 @@ export class DatasetValidator {
       if (control.value == 0) {
         return { msg: { value: 'Please set data source!' } };
       }
-      return null;
+      return null!;
     };
   }
 
-  static localFile(projectType, enableAWSS3, duplicate): ValidatorFn {
+  static localFile(projectType: any, enableAWSS3: any, duplicate: any): ValidatorFn {
     return (control: AbstractControl): { [key: string]: any } => {
       if (!control.parent) {
-        return null;
+        return null!;
       }
-      const inputFile = control.parent.get('localFile').value;
+      const inputFile = control.parent?.get('localFile')?.value;
       if (inputFile) {
         if (!DatasetValidator.validateFileExt(inputFile, projectType)) {
           return { msg: { value: DatasetValidator.FILE_FORMAT_NOT_SUPPORT } };
@@ -160,33 +146,49 @@ export class DatasetValidator {
           return { msg: { value: DatasetValidator.FILE_SIZE_EXCEED_LIMIT } };
         }
         if (!enableAWSS3) {
-          return duplicate ? { msg: { value: DatasetValidator.FILE_DUPLICATED } } : null;
+          return duplicate ? { msg: { value: DatasetValidator.FILE_DUPLICATED } } : null!;
+        }
+      }
+    };
+  }
+
+  static localLabelFile(type: string, inputFile: any): ValidatorFn {
+    return (control: AbstractControl): { [key: string]: any } => {
+      if (!control.parent) {
+        return null!;
+      }
+      if (inputFile) {
+        if (!DatasetValidator.validateLabelFile(type, inputFile)) {
+          return { msg: { value: DatasetValidator.FILE_FORMAT_NOT_SUPPORT } };
+        }
+        if (!DatasetValidator.validateFileSize(inputFile)) {
+          return { msg: { value: DatasetValidator.FILE_SIZE_EXCEED_LIMIT } };
         }
       }
       return DatasetValidator.isEmpty(control.value)
         ? { msg: { value: DatasetValidator.REQUIRED_FIELD } }
-        : null;
+        : (null as any);
     };
   }
 
   static imageFile(): ValidatorFn {
     return (control: AbstractControl): { [key: string]: any } => {
       if (!control.parent) {
-        return null;
+        return null!;
       }
 
-      const inputFile = control.parent.get('localFile').value;
+      const inputFile = control.parent!.get('localFile')!.value;
       if (inputFile) {
         if (!DatasetValidator.validateImageFile(inputFile)) {
           return { msg: { value: DatasetValidator.FILE_FORMAT_NOT_SUPPORT } };
         }
         if (!DatasetValidator.validateImageFileSize(inputFile)) {
-          return { msg: { value: DatasetValidator.IMAGE_FILE_SIZE_EXCEED_LIMIT } };
+          return {
+            msg: { value: DatasetValidator.IMAGE_FILE_SIZE_EXCEED_LIMIT },
+          };
         }
       }
-      return DatasetValidator.isEmpty(control.value)
-        ? { msg: { value: DatasetValidator.REQUIRED_FIELD } }
-        : null;
+      return DatasetValidator.isEmpty(control.value) ? { msg: { value: DatasetValidator.REQUIRED_FIELD } } : null!;
     };
   }
 
@@ -195,11 +197,11 @@ export class DatasetValidator {
       if (DatasetValidator.isEmpty(control.value)) {
         return { msg: { value: DatasetValidator.REQUIRED_FIELD } };
       }
-      return null;
+      return null!;
     };
   }
 
-  static requiredTwo(type): ValidatorFn {
+  static requiredTwo(type: string): ValidatorFn {
     return (control: AbstractControl): { [key: string]: any } => {
       if (type == 'ner' || type == 'image' || type == 'log') {
         if (control.value.length < 1) {
@@ -210,33 +212,25 @@ export class DatasetValidator {
           return { msg: { value: DatasetValidator.REQUIRED_FIELD_LABEL } };
         }
       }
-      return null;
+      return null!;
     };
   }
 
-  static requiredTwoPopLabel(type): ValidatorFn {
+  static requiredTwoPopLabel(type: string): ValidatorFn {
     return (control: AbstractControl): { [key: string]: any } => {
       if (type == 'ner') {
         if (control.value.length < 2) {
           return { msg: { value: DatasetValidator.REQUIRED_FIELD_POP_LABEL } };
         }
       }
-      return null;
-    };
-  }
-
-  static sizeLimit(): ValidatorFn {
-    return (control: AbstractControl): { [key: string]: any } => {
-      return DatasetValidator.isEmpty(control.value)
-        ? { msg: { value: DatasetValidator.REQUIRED_FIELD } }
-        : null;
+      return null!;
     };
   }
 
   static numberOnly(): ValidatorFn {
     return (control: AbstractControl): { [key: string]: any } => {
       if (!control.parent) {
-        return null;
+        return null!;
       }
 
       if (DatasetValidator.isEmpty(control.value)) {
@@ -248,14 +242,14 @@ export class DatasetValidator {
       if (!String(control.value).match(argRegEx)) {
         return { msg: { value: 'Wrong format! number only!' } };
       }
-      return null;
+      return null!;
     };
   }
 
   static maxAnnotation(): ValidatorFn {
     return (control: AbstractControl): { [key: string]: any } => {
       if (!control.parent) {
-        return null;
+        return null!;
       }
       const pattern = `^[0-9]+$`;
       const argRegEx = new RegExp(pattern, 'g');
@@ -265,16 +259,18 @@ export class DatasetValidator {
         control.value == null ||
         control.value == undefined
       ) {
-        return { msg: { value: 'Please enter a integer number greater than 0.' } };
+        return {
+          msg: { value: 'Please enter a integer number greater than 0.' },
+        };
       }
-      return null;
+      return null!;
     };
   }
 
   static threshold(): ValidatorFn {
     return (control: AbstractControl): { [key: string]: any } => {
       if (!control.parent) {
-        return null;
+        return null!;
       }
 
       if (DatasetValidator.isEmpty(control.value)) {
@@ -288,37 +284,19 @@ export class DatasetValidator {
       }
 
       if (control.value < 50) {
-        return { msg: { value: 'Should not less than the minimum threshold 50 !' } };
+        return {
+          msg: { value: 'Should not less than the minimum threshold 50 !' },
+        };
       }
 
-      return null;
-    };
-  }
-  static minNumber(): ValidatorFn {
-    return (control: AbstractControl): { [key: string]: any } => {
-      if (!control.parent) {
-        return null;
-      }
-
-      if (control.value == null || control.value == undefined) {
-        return { msg: { value: DatasetValidator.REQUIRED_FIELD } };
-      }
-
-      const pattern = `^(\\-|\\+)?\\d+(\\.\\d+)?$`;
-      // let pattern = `^[0-9]+$`;
-
-      const argRegEx = new RegExp(pattern, 'g');
-      if (!String(control.value).match(argRegEx)) {
-        return { msg: { value: 'Wrong format! number only' } };
-      }
-      return null;
+      return null!;
     };
   }
 
   static textOnly(): ValidatorFn {
     return (control: AbstractControl): { [key: string]: any } => {
       if (!control.parent) {
-        return null;
+        return null!;
       }
 
       if (DatasetValidator.isEmpty(control.value)) {
@@ -330,11 +308,11 @@ export class DatasetValidator {
       if (!control.value.match(argRegEx)) {
         return { msg: { value: 'Wrong format! text only' } };
       }
-      return null;
+      return null!;
     };
   }
 
-  private static validateFileExt(inputFile: any, projectType): boolean {
+  private static validateFileExt(inputFile: any, projectType: string): boolean | undefined {
     const ext = inputFile.name.split('.').pop().toLowerCase();
     if (projectType || projectType == '') {
       if (projectType == 'image') {
@@ -367,6 +345,16 @@ export class DatasetValidator {
     } else {
       return false;
     }
+    return false;
+  }
+
+  private static validateLabelFile(type: string, inputFile: any): boolean {
+    const ext = inputFile.name.split('.').pop().toLowerCase();
+    if (type === 'json') {
+      return type === ext.toLowerCase();
+    } else {
+      return ['yaml', 'yml'].includes(ext.toLowerCase());
+    }
   }
 
   private static validateImageFile(inputFile: any): boolean {
@@ -392,17 +380,9 @@ export class DatasetValidator {
   private static validateFileSize(inputFile: any): boolean {
     const size = inputFile.size;
     return true;
-
-    // 500 MB
-    // if (inputFile.size < 524288000) {
-    //   return true;
-    // }
-    // else {
-    //   return false;
-    // }
   }
 
-  private static isEmpty(str): boolean {
+  private static isEmpty(str: any): boolean {
     if (!str) {
       return true;
     } else if (str instanceof String && str.trim().length === 0) {
@@ -415,7 +395,7 @@ export class DatasetValidator {
     return false;
   }
 
-  private static isEmptyObject(obj): boolean {
+  private static isEmptyObject(obj: any): boolean {
     for (const key in obj) {
       if (obj[key] !== null && obj[key] !== '') {
         return false;
@@ -424,22 +404,46 @@ export class DatasetValidator {
     return true;
   }
 
-  static isInvalidNumber(input) {
-    return (input === null || input === '');
+  static isInvalidNumber(input: any) {
+    return input === null || input === '';
   }
 
-  static isNotIntegerNum(input) {
+  static isNotIntegerNum(input: any) {
     return Math.round(input) !== input;
   }
 
-  static isRepeatArr(arr) { 
-    let hasArr = {}; 
-    for (var i in arr) { 
+  static isRepeatArr(arr: any) {
+    let hasArr: any = {};
+    for (var i in arr) {
       if (arr[i] && hasArr[arr[i]]) {
-        return true; 
+        return true;
       }
-      hasArr[arr[i]] = true; 
-    } 
-    return false; 
+      hasArr[arr[i]] = true;
+    }
+    return false;
+  }
+
+  static uploadFileSizeLimit(uploadFrom, configFileSize): ValidatorFn {
+    return (control: AbstractControl): { [key: string]: any } => {
+      if (!control.parent) {
+        return null;
+      }
+      const inputFile = control.parent.get('localFile').value;
+      if (inputFile) {
+        if (inputFile.size > 104857600 && uploadFrom == 'create') {
+          return { msg: { value: DatasetValidator.FILE_EXCEEDS_100MB } };
+        } else if (inputFile.size > configFileSize && uploadFrom == 'datasets') {
+          const limitFileSize =
+            configFileSize >= 1024 * 1024 * 1024
+              ? configFileSize / (1024 * 1024 * 1024) + 'GB'
+              : configFileSize / (1024 * 1024) + 'MB';
+          return {
+            msg: {
+              value: `File size exceeds the maximum ${limitFileSize}. Please select a new file to upload. `,
+            },
+          };
+        }
+      }
+    };
   }
 }
