@@ -1,11 +1,15 @@
-import insertAfter from "./index.es5";
+/*
+Copyright 2019-2023 VMware, Inc.
+SPDX-License-Identifier: Apache-2.0
+*/
+import insertAfter from './index.es5';
 
-function isTextNode(node) {
+function isTextNode(node: any) {
   const TEXT_NODE = 3;
   return node.nodeType === TEXT_NODE;
 }
 
-function getNextNode(node) {
+function getNextNode(node: any) {
   if (node.firstChild) return node.firstChild;
   while (node) {
     if (node.nextSibling) return node.nextSibling;
@@ -13,7 +17,7 @@ function getNextNode(node) {
   }
 }
 
-function getNodesInRange(range) {
+function getNodesInRange(range: any) {
   var start = range.startContainer;
   var end = range.endContainer;
   var commonAncestor = range.commonAncestorContainer;
@@ -36,20 +40,20 @@ function getNodesInRange(range) {
   return nodes;
 }
 
-function splitText(node, offset) {
+function splitText(node: any, offset: any) {
   let tail = node.cloneNode(false);
   tail.deleteData(0, offset);
   node.deleteData(offset, node.length - offset);
   return insertAfter(tail, node);
 }
 
-function highlightRange(normedRange, cssClass, cssStyle) {
+function highlightRange(normedRange: any, cssClass: any, cssStyle: any) {
   if (typeof cssClass === 'undefined' || cssClass === null) {
     cssClass = 'spanMarked';
   }
 
   const allNodes = getNodesInRange(normedRange._range);
-  const textNodes = allNodes.filter(n => isTextNode(n));
+  const textNodes = allNodes.filter((n) => isTextNode(n));
 
   var white = /^\s*$/;
 
@@ -62,7 +66,7 @@ function highlightRange(normedRange, cssClass, cssStyle) {
 
   let nlen = nodes.length;
   if (nlen > 1 && nodes[nodes.length - 1].length !== normedRange._range.endOffset) {
-     nlen = nlen - 1;
+    nlen = nlen - 1;
   }
   const results = [];
   for (var i = start, len = nlen; i < len; i++) {
@@ -79,7 +83,7 @@ function highlightRange(normedRange, cssClass, cssStyle) {
   return results;
 }
 
-function splitBoundaries(range) {
+function splitBoundaries(range: any) {
   let { startContainer, startOffset, endContainer, endOffset } = range;
 
   if (isTextNode(endContainer)) {
@@ -102,14 +106,14 @@ function splitBoundaries(range) {
   }
 }
 
-const toGlobalOffset = (container, element, len) => {
+const toGlobalOffset = (container: any, element: any, len: any) => {
   let pos = 0;
-  const count = node => {
+  const count = (node: any) => {
     if (node === element) {
       return pos;
     }
-    if (node.nodeName === "#text") pos = pos + node.length;
-    if (node.nodeName === "BR") pos = pos + 1;
+    if (node.nodeName === '#text') pos = pos + node.length;
+    if (node.nodeName === 'BR') pos = pos + 1;
 
     for (var i = 0; i <= node.childNodes.length; i++) {
       const n = node.childNodes[i];
@@ -123,32 +127,26 @@ const toGlobalOffset = (container, element, len) => {
   return len + count(container);
 };
 
-function removeSpans(spans) {
+function removeSpans(spans: any) {
   var norm = [];
 
   if (spans) {
-    spans.forEach(span => {
+    spans.forEach((span: any) => {
       while (span.firstChild) span.parentNode.insertBefore(span.firstChild, span);
       norm.push(span.parentNode);
       span.parentNode.removeChild(span);
     });
   }
 
-  norm.forEach(n => n.normalize());
+  norm.forEach((n) => n.normalize());
 }
 
-function findClosestTextNode(node, lookBack = false) {
+function findClosestTextNode(node: any, lookBack = false) {
   if (!node) return null;
   if (node.nodeName === '#text') return node;
   const child = findClosestTextNode(lookBack ? node.lastChild : node.firstChild, lookBack);
   if (child) return child;
   return findClosestTextNode(lookBack ? node.previousSibling : node.nextSibling, lookBack);
-};
+}
 
-export {
-  removeSpans,
-  toGlobalOffset,
-  highlightRange,
-  splitBoundaries,
-  findClosestTextNode,
-};
+export { removeSpans, toGlobalOffset, highlightRange, splitBoundaries, findClosestTextNode };
