@@ -1,10 +1,10 @@
 /*
-Copyright 2019-2021 VMware, Inc.
+Copyright 2019-2023 VMware, Inc.
 SPDX-License-Identifier: Apache-2.0
 */
-import { Env } from 'app/model/index';
 
-export const environment:Env = {
-  production: true
+import { Env } from 'src/app/model/index';
+
+export const environment: Env = {
+  production: true,
 };
-
