@@ -1,5 +1,5 @@
 /*
-Copyright 2019-2021 VMware, Inc.
+Copyright 2019-2023 VMware, Inc.
 SPDX-License-Identifier: Apache-2.0
 */
 
@@ -29,8 +29,8 @@ function userChart(options) {
     .attr('height', height + margin.top + margin.bottom)
     .call(zoom);
 
-  let tip = d3
-    .tip()
+  tip = options.tip;
+  tip
     .attr('class', 'd3-tip')
     .attr('class', 'd3-tip-user')
     .style('color', 'white')
@@ -39,7 +39,7 @@ function userChart(options) {
     .style('border-radius', '4px')
     .style('font-size', '10px')
     .offset([-10, 0])
-    .html(function (d) {
+    .html(function (event, d) {
       return '<span>' + d.value + '</span>';
     });
 
@@ -82,11 +82,11 @@ function userChart(options) {
     .attr('y', (d) => y(d.value))
     .attr('height', (d) => y(0) - y(d.value))
     .attr('width', d3.min([x.bandwidth(), 150]))
-    .on('mouseover', function (d) {
-      tip.show(d, this);
+    .on('mouseover', function (event, d) {
+      tip.show(event, d);
       d3.select(this).style('opacity', 0.5);
     })
-    .on('mouseout', function (d) {
+    .on('mouseout', function (event, d) {
       tip.hide();
       d3.select(this).style('opacity', 1);
     });
@@ -106,8 +106,8 @@ function userChart(options) {
       d3.zoom().scaleExtent([1, 5]).translateExtent(extent).extent(extent).on('zoom', zoomed),
     );
 
-    function zoomed() {
-      x.range([margin.left, width - margin.right].map((d) => d3.event.transform.applyX(d)));
+    function zoomed(event, d) {
+      x.range([margin.left, width - margin.right].map((d) => event.transform.applyX(d)));
       svg
         .selectAll('.bars rect')
         .attr('x', (d) => x(d.name))
