@@ -1,13 +1,13 @@
 /*
-Copyright 2019-2021 VMware, Inc.
+Copyright 2019-2023 VMware, Inc.
 SPDX-License-Identifier: Apache-2.0
 */
 
-import './polyfills.ts';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
-import { enableProdMode } from '@angular/core';
+import { DevPlatformModule } from './dev-platform/dev-platform.module';
 import { environment } from './environments/environment';
-import { AppModule } from './app/';
+import { MicroFrontendModule } from './micro-frontend/micro-frontend.module';
+import { enableProdMode } from '@angular/core';
 
 if (environment.production && environment.googleTrackId) {
   enableProdMode();
@@ -22,4 +22,8 @@ if (environment.production && environment.googleTrackId) {
   document.head.appendChild(script2);
 }
 
-platformBrowserDynamic().bootstrapModule(AppModule);
+const bootstrapModule = environment.embedded ? MicroFrontendModule : DevPlatformModule;
+
+platformBrowserDynamic()
+  .bootstrapModule(bootstrapModule)
+  .catch((err) => console.error('platformBrowserDynamic:::', err));
