@@ -6,14 +6,14 @@
 ***/
 
 const config = require('../config/config');
-const { USER_ROLE } = require('../config/constant');
+const { USER_ROLE, GENERATESTATUS, APPENDSR, QUERYORDER, ANNOTATION_QUESTION, TICKET_DESCRIPTION } = require('../config/constant');
 const mongoose = require("mongoose"),
     Schema = mongoose.Schema;
 const mongoosePaginate = require('mongoose-paginate-v2');
 
 // MongoDB database name
 connectOptions = { autoIndex: config.mongoDBAutoIndex, useNewUrlParser: true, useUnifiedTopology: true };
-if(config.ESP){
+if (config.ESP) {
     connectOptions = { autoIndex: config.mongoDBAutoIndex };
 }
 
@@ -41,16 +41,22 @@ const srSchema = new mongoose.Schema({
         users: [],
         silence: { type: Boolean, default: false }
     },
+    skip: {
+        users: []
+    },
     text_vector: { type: String },
     al_test: { type: Boolean },
-    reviewInfo:{
-        user: { type: String },
-        reviewed: { type: Boolean, default: false },
+    reviewInfo: {
+        userInputs: [],
         review: { type: Boolean, default: false },
+        reviewed: { type: Boolean, default: false },
+        modified: { type: Boolean, default: false },
+        passed: { type: Boolean, default: false },
+        user: { type: String },
         reviewedTime: { type: String },
-        modified: { type: Boolean, default: false }
     },
     ticketQuestions: { type: Object },
+    questionForText:{ type: Array, default: []}
 }, { _id: true });
 srSchema.set("toJSON", { virtuals: true });
 srSchema.index({ projectName: 1 });
@@ -73,12 +79,17 @@ const imgSchema = new mongoose.Schema({
         users: [],
         silence: { type: Boolean, default: false }
     },
-    reviewInfo:{
-        user: { type: String },
-        reviewed: { type: Boolean, default: false },
+    skip: {
+        users: []
+    },
+    reviewInfo: {
+        userInputs: [],
         review: { type: Boolean, default: false },
+        reviewed: { type: Boolean, default: false },
+        modified: { type: Boolean, default: false },
+        passed: { type: Boolean, default: false },
+        user: { type: String },
         reviewedTime: { type: String },
-        modified: { type: Boolean, default: false }
     }
 }, { _id: true });
 imgSchema.set("toJSON", { virtuals: true });
@@ -97,17 +108,22 @@ const logSchema = new mongoose.Schema({
     userInputsLength: { type: Number },
     userInputs: [],
     originalData: { type: Object },
-    fileInfo:{ type: Object },
+    fileInfo: { type: Object },
     flag: {
         users: [],
         silence: { type: Boolean, default: false }
     },
-    reviewInfo:{
-        user: { type: String },
-        reviewed: { type: Boolean, default: false },
+    skip: {
+        users: [],
+    },
+    reviewInfo: {
+        userInputs: [],
         review: { type: Boolean, default: false },
+        reviewed: { type: Boolean, default: false },
+        modified: { type: Boolean, default: false },
+        passed: { type: Boolean, default: false },
+        user: { type: String },
         reviewedTime: { type: String },
-        modified: { type: Boolean, default: false }
     }
 }, { _id: true });
 logSchema.set("toJSON", { virtuals: true });
@@ -124,14 +140,23 @@ const LogModel = mongoose.model("LOG", logSchema);
 const userSchema = new mongoose.Schema({
     _id: { type: String },
     email: { type: String },
-    password: {type: String},
+    password: { type: String },
     fullName: { type: String },
     points: { type: Number, default: 0 },
     role: { type: String, default: USER_ROLE },
-    createdDate: { type: String, default: Date.now() }
+    createdDate: { type: String },
+    updateDate: { type: String },
+    regularNotification: { type: Boolean, default: true },
+    manul: { type: Boolean, default: false },
+    dashboard:{
+        createdDate: { type: String },
+        updateDate: { type: String },
+        data:[],
+    },
 }, { _id: false });
 userSchema.set("toJSON", { virtuals: true });
 userSchema.set("toObject", { virtuals: true });
+userSchema.plugin(mongoosePaginate);
 
 // create User model
 const UserModel = mongoose.model("User", userSchema);
@@ -145,42 +170,43 @@ const projectSchema = new mongoose.Schema({
     createdDate: { type: String },
     updatedDate: { type: String },
     projectName: { type: String },
-    taskInstructions: { type: String },
+    taskInstructions: { type: String, default: "" },
     totalCase: { type: Number },
-    projectCompleteCase: { type: Number },
+    projectCompleteCase: { type: Number, default: 0 },
     userCompleteCase: [{
         user: { type: String },
         completeCase: { type: Number, default: 0 },
-        skip: { type: Number, default: 0 },
         reviewed: { type: Number, default: 0 },
         assignedCase: { type: Number },
+        assignedDate: { type: String },
+        updateDate: { type: String },
+        regularNotification: { type: Boolean, default: true },
     }],
-    reviewInfo:[{
+    reviewInfo: [{
         user: { type: String },
         reviewedCase: { type: Number, default: 0 },
-        skip: { type: Number, default: 0 },
     }],
-    maxAnnotation: { type: Number },
+    maxAnnotation: { type: Number, default: 1 },
     categoryList: { type: String },
-    assignmentLogic: { type: String },
+    assignmentLogic: { type: String, default: QUERYORDER.RANDOM },
     annotator: { type: Array },
-    dataSource: { type: String },
+    dataSource: { type: String, default: "" },
     selectedDataset: { type: Array },
     selectedColumn: { type: Array },
-    annotationQuestion: { type: String, default: 'What label does this ticket belong to ?' },
-    shareStatus: { type: Boolean },
-    shareDescription: { type: String },
+    annotationQuestion: { type: String, default: ANNOTATION_QUESTION },
+    shareStatus: { type: Boolean, default: false },
+    shareDescription: { type: String, default: "" },
     generateInfo: {
-        status: { type: String, default: "pending" },
+        status: { type: String, default: GENERATESTATUS.DEFAULT },
         messageId: { type: String },
         startTime: { type: String },
         updateTime: { type: String },
         file: { type: String },
         format: { type: String, default: "standard" },
-        onlyLabelled: { type: String}
+        onlyLabelled: { type: String }
     },
-    fileSize: { type: Number, default: 0 },
-    appendSr: { type: String, default: "pending" },
+    fileSize: { type: Number, default: 1 },
+    appendSr: { type: String, default: APPENDSR.DEFAULT },
     labelType: { type: String },
     projectType: { type: String },
     min: { type: Number, default: 0 },
@@ -193,7 +219,7 @@ const projectSchema = new mongoose.Schema({
         model: { type: String },
         vectorModel: { type: String },
         newLBSr: { type: Array },
-        queriedSr: [ { type: Schema.Types.ObjectId } ],
+        queriedSr: [{ type: Schema.Types.ObjectId }],
         frequency: { type: Number, default: 10 },
         trigger: { type: Number, default: 50 },
         trained: { type: Boolean, default: false },
@@ -209,14 +235,19 @@ const projectSchema = new mongoose.Schema({
     },
     encoder: { type: String },
     isMultipleLabel: { type: Boolean, default: false },
-    regression: {type: Boolean, default: false},
-    isShowFilename: {type: Boolean, default: false},
-    ticketDescription: {type: String, default: 'Passage'},
-    ticketQuestions: { type: Array },
-    popUpLabels:{ type: Array },
+    regression: { type: Boolean, default: false },
+    isShowFilename: { type: Boolean, default: false },
+    ticketDescription: { type: String, default: TICKET_DESCRIPTION },
+    popUpLabels: { type: Array },
+    integration: {
+        source: { type: String, default: "" },
+        externalId: { type: Array, default: [] },
+    },
+    assignSlackChannels: { type: Array },
 }, { _id: true });
 projectSchema.index({ projectName: 1 });
 projectSchema.set("toJSON", { virtuals: true });
+projectSchema.plugin(mongoosePaginate);
 
 // create Annotation Project model
 const ProjectModel = mongoose.model("Project", projectSchema);
@@ -240,18 +271,33 @@ const dataSetSchema = new mongoose.Schema({
         type: { type: String },
         uniqueLength: { type: Number }
     }],
-    images:[{
+    images: [{
         fileName: { type: String },
         location: { type: String },
         fileSize: { type: Number },
     }],
-    totalRows: { type: Number }
+    totalRows: { type: Number },
+    totalColumns: { type: Number },
+    dataSynchronize:[{
+        system: { type: String },
+        _id: { type: String },
+    }],
+    projects:{ type: Array },
 });
 dataSetSchema.set("toJSON", { virtuals: true });
 dataSetSchema.index({ dataSetName: 1 });
+dataSetSchema.plugin(mongoosePaginate);
 
 const DataSetModel = mongoose.model("DataSet", dataSetSchema);
 
+
+//dataSet Model
+const instanceSchema = new mongoose.Schema({
+    data: { type: String, required: true, unique: true },
+});
+
+const InstanceModel = mongoose.model("instance", instanceSchema);
+InstanceModel.createIndexes().catch(err => console.log('[ DB-instance ]', err));
 
 module.exports = {
     SrModel,
@@ -260,6 +306,7 @@ module.exports = {
     UserModel,
     ProjectModel,
     DataSetModel,
+    InstanceModel,
     db,
 
 }
