@@ -10,6 +10,8 @@ const express = require("express");
 const router = express.Router();
 const emailService = require('../services/email-service');
 const APIs = require('../resources/APIs');
+const taskSchedule = require('../utils/taskSchedule');
+const config = require('../config/config');
 
 router.post(APIs.EMAIL_TO_OWNER, (req, res) => {
   console.log(`[ EMAIL ] [ ACCESS ] Router ${req.originalUrl} ${req.auth.email}`);
@@ -18,7 +20,7 @@ router.post(APIs.EMAIL_TO_OWNER, (req, res) => {
     res.status(200).json("email ok");
   }).catch(error => {
     console.error(`[ EMAIL ] [ ERROR ] Router ${req.originalUrl} ${req.auth.email}`, error);
-    res.status(500).json("email failed");
+    res.status(500).json(error);
   });
 });
 
@@ -30,7 +32,18 @@ router.post(APIs.EMAIL_TO_ANNOTATOR, (req, res) => {
     res.status(200).json("email ok");
   }).catch(error => {
     console.error(`[ EMAIL ] [ ERROR ] Router ${req.originalUrl} ${req.auth.email}`, error);
-    res.status(500).json("email failed");
+    res.status(500).json(error);
+  });
+});
+
+router.get(APIs.EMAIL_REGULAR_NOTIFICATION, (req, res) => {
+  console.log(`[ EMAIL ] [ ACCESS ] Router ${req.originalUrl} ${req.query.u}`);
+  taskSchedule.regularNotificationSubscription(req).then(() => {
+    console.log(`[ EMAIL ] [ SUCCESS ] Router ${req.originalUrl} ${req.query.u}`);
+    res.status(200).redirect(`${config.WebClientUrl}/home?o=email&s=1`);
+  }).catch(error => {
+    console.error(`[ EMAIL ] [ ERROR ] Router ${req.originalUrl} ${req.query.u}`, error);
+    res.status(500).redirect(`${config.WebClientUrl}/home?o=email&s=0`);
   });
 });
 
