@@ -11,22 +11,30 @@ const path = require('path');
 const http = require('http');
 const bodyParser = require('body-parser');
 const cors = require('cors');
-const glob = require('glob');
+const fs = require('fs')
 const { consumeSQSMessage } = require('./utils/sqs');
+const { regularNotification } = require('./utils/taskSchedule');
 const config = require('./config/config');
-const { API_VERSION } = require('./config/constant')
+const { API_VERSION, API_BASE_PATH } = require('./config/constant');
+const MESSAGE = require('./config/code_msg');
 
 // Get our API routes
 const authService = require('./services/auth.service');
-const {jwtTokenAuthrization} = require('./middlewares/jwt.middleware');
+const { jwtTokenAuthrization } = require('./middlewares/jwt.middleware');
 
 const app = express();
-const routers = glob.sync(path.resolve(__dirname, `./routers/*.js`));
+
+const routers = fs.readdirSync(path.join(__dirname, 'routers'));
 
 //manually swagger
 const swaggerUi = require('swagger-ui-express');
 const swaggerDocument = require('./swagger2.json');
 const ua = require('universal-analytics');
+
+
+//slack
+const slackService = require('./services/slack/slack')
+
 
 // Parsers for POST data
 app.use(bodyParser.json());
@@ -43,11 +51,15 @@ app.use(function (req, res, next) {
 });
 
 // mannually set up swagger
-app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));
+app.use(`${API_BASE_PATH}/api-docs`, swaggerUi.serve, swaggerUi.setup(swaggerDocument));
 
 // Server health check
-app.get('/health', (req, res) => {
-  return res.status(200).json({CODE: 200, MSG: "SUCCESS"});
+app.get(`${API_BASE_PATH}/health`, (req, res) => {
+  return res.status(200).json(MESSAGE.SUCCESS);
+});
+
+process.on('uncaughtException', err => {
+  console.error('[SERVER-ERROR]', err);
 });
 
 // esp author
@@ -57,30 +69,31 @@ authService.authentication().then(data => {
   if (config.trackingId) {
     app.use(function (req, res, next) {
       if (req.headers.referer && req.headers.referer.includes("api-docs")) {
-        const visitor = ua(config.trackingId, {uid: req.auth.email});
+        const visitor = ua(config.trackingId, { uid: req.auth.email });
         visitor.pageview(req.originalUrl).send();
       }
       next();
     });
   }
-   
+
   app.use(function (err, req, res, next) {
     if (err.name === 'UnauthorizedError') {
-      res.status(401).send({ MSG: "Invalid token" });
+      res.status(401).send(MESSAGE.VALIDATION_PERMITION);
     }
   });
 }).catch(error => {
   console.error(`[ SERVER-AUTH ][ ERROR ]`, error.message);
-}).finally(()=>{
-    // Set our api routers
-    routers.forEach(
-      api => app.use(`/api/${API_VERSION}`, require(api))
-    );
-    if (config.ESP || config.useAWS &&  config.sqsRoleArn && config.sqsUrl) {
-      //consume SQS message
-      consumeSQSMessage();
-    }
-
-    const server = http.createServer(app);
-    server.listen(config.serverPort, () => console.log(`[ SERVER ] API running on localhost:${config.serverPort}`));
+}).finally(() => {
+  // Set our api routers
+  routers.forEach(
+    api => app.use(`${API_BASE_PATH}/api/${API_VERSION}`, require(`./routers/${api}`))
+  );
+  consumeSQSMessage();
+  regularNotification();
+  if (config.buildSlackApp) {
+    slackService.slackStart();
+  }
+  const server = http.createServer(app);
+  server.listen(config.serverPort, () => console.log(`[ SERVER ] API running on localhost:${config.serverPort}`));
 });
+
