@@ -11,11 +11,12 @@ const config = require('../config/config');
 const { generateBasicToken } = require('../middlewares/jwt.middleware');
 const userService = require('./user-service');
 const https = require('https');
+const MESSAGE = require('../config/code_msg')
 
 let publicKey;
 
-const obtainPublicKey = async() => {
-    const response = await axios.create({validateStatus: () => { return true }}).get(`${config.authServiceUrl}/api/auth/v1/tokens/public-key`);
+const obtainPublicKey = async () => {
+    const response = await axios.create({ validateStatus: () => { return true } }).get(`${config.authServiceUrl}/api/auth/v1/tokens/public-key`);
 
     if (response.status === 200) {
         console.log(`[ AUTH ] Service successfully obtainPublicKey`);
@@ -30,7 +31,7 @@ const obtainPublicKey = async() => {
     throw new Error(error.msg, error.data);
 };
 
-async function getPublicKey(){
+async function getPublicKey() {
     console.log(`[ AUTH ] Service get public key`);
     if (publicKey === undefined) {
         publicKey = await obtainPublicKey();
@@ -46,17 +47,17 @@ async function authentication() {
 }
 
 
-async function login(req){
+async function login(req) {
     if (!req.body.email || !req.body.password) {
-        throw {CODE: 401, MSG: "USERNAME OR PASSWORD IS EMPTY"};
+        throw MESSAGE.VALIDATION_UNAUTH;
     }
     let user;
     if (req.body.ldap) {
         user = await authenticateWithLDAP(req);
-    }else{
+    } else {
         user = await basicLogin(req);
     }
-    
+
     const token = await generateBasicToken(user.email);
     return {
         token: token,
@@ -70,39 +71,40 @@ async function refreshToken(req) {
     return generateBasicToken(req.auth.email);
 }
 
-async function basicLogin(req){
+async function basicLogin(req) {
     const user = await userService.queryUserById(req.body.email);
     if (!user) {
-        throw {CODE: 401, MSG: "USERNAME OR PASSWORD IS INVALID"};
+        throw MESSAGE.VALIDATION_UNAUTH;
     }
     if (user.password != Buffer.from(req.body.password).toString("base64")) {
-        throw {CODE: 401, MSG: "USERNAME OR PASSWORD IS INVALID"};
+        throw MESSAGE.VALIDATION_UNAUTH;
     }
     return user;
 }
 
 async function authenticateWithLDAP(req) {
     if (!config.loginWithLDAP) {
-        throw {CODE: 4002, MSG: "MISSING THE LDAP AUTHORIZATION LINK"};
+        throw MESSAGE.VALIDATION_LDAP;
     }
     const userBtoa = Buffer.from(`${req.body.email}:${req.body.password}`).toString('base64');
     const requestOptions = {
         method: 'post',
         url: config.loginWithLDAP,
-        headers: {"Authorization": "Basic " + userBtoa },
-        httpsAgent: new https.Agent({rejectUnauthorized: false})
+        headers: { "Authorization": "Basic " + userBtoa },
+        httpsAgent: new https.Agent({ rejectUnauthorized: false })
     };
     try {
-        resp =  await axios.request(requestOptions);
+        resp = await axios.request(requestOptions);
         if (resp.status != 200) {
-            throw {CODE: 401, MSG: "USERNAME OR PASSWORD IS INVALID"};
+            throw MESSAGE.VALIDATION_UNAUTH;
         }
-        const email = resp.data.emailAddress? resp.data.emailAddress: resp.data.email;
-        const userName = resp.data.userName? resp.data.userName: resp.data.fullName;
+        const email = resp.data.emailAddress ? resp.data.emailAddress : resp.data.email;
+        const userName = resp.data.userName ? resp.data.userName : resp.data.fullName;
         return userService.queryAndUpdateUser(email, userName);
 
     } catch (error) {
-        throw {CODE: error.code, MSG: error.message};
+        MESSAGE.ERROR.DATA = [{ CODE: error.code, MSG: error.message }]
+        throw MESSAGE.ERROR;
     }
 }
 
