@@ -17,6 +17,8 @@ const S3Utils = require('../utils/s3');
 const localFileSysService = require('./localFileSys.service');
 const mongoDb = require('../db/mongo.db');
 const { ProjectModel } = require('../db/db-connect');
+const validator = require('../utils/validator');
+const MESSAGE = require('../config/code_msg');
 
 async function generateFile(req){
     
@@ -30,11 +32,12 @@ async function generateFile(req){
     let response = {};
     
     console.log(`[ SQS ] Service generateFile query file size and generate info`);
-    const pro = await mongoDb.findById(ProjectModel, ObjectId(data.id));
+    let project_check = await validator.checkProjectByconditions({ _id: data.id }, true);
+    const pro = project_check[0];
     const gen = pro.generateInfo;
 
-    if ((pro.projectType == PROJECTTYPE.NER || pro.projectType == PROJECTTYPE.IMGAGE) && data.format != FILEFORMAT.STANDARD){
-        return {CODE: 4001, MSG: "ERROR FORMAT"};
+    if ((pro.projectType == PROJECTTYPE.NER || pro.projectType == PROJECTTYPE.QA || pro.projectType == PROJECTTYPE.IMGAGE) && data.format != FILEFORMAT.STANDARD){
+        return MESSAGE.VALIDATATION_PJ_FORMAT;
     }
     
     if (config.useLocalFileSys) {
