@@ -17,20 +17,20 @@ async function prepareCredentials(resource, sessionDuration) {
     console.log(`[ STS ] Utils prepareCredentials`);
     if (config.ESP) {
         return prepareESPCredentails(resource, sessionDuration);
-    }else{
+    } else {
         return prepareOSCredentails(resource, sessionDuration);
     }
 }
 
-async function prepareOSCredentails(resource, sessionDuration){
+async function prepareOSCredentails(resource, sessionDuration) {
 
     let roleArn = "";
     if (resource === AWSRESOURCE.S3) {
         roleArn = config.s3RoleArn;
-    } else if (resource === AWSRESOURCE.SQS){
+    } else if (resource === AWSRESOURCE.SQS) {
         roleArn = config.sqsRoleArn;
     }
-    
+
     let sts = new AWS.STS({
         endpoint: `https://sts.${config.region}.amazonaws.com`,
         apiVersion: "2011-06-15",
@@ -38,17 +38,17 @@ async function prepareOSCredentails(resource, sessionDuration){
         accessKeyId: config.accessKeyId,
         secretAccessKey: config.secretAccessKey
     });
-    
+
     const stsAssecume = {
         DurationSeconds: sessionDuration,//defalut 1H min 15min
         RoleSessionName: "loop", //can be any value
         RoleArn: roleArn
     };
     console.log(`[ STS ] Utils sts.assumeRole`);
-    return await sts.assumeRole(stsAssecume).promise();
+    return sts.assumeRole(stsAssecume).promise();
 }
 
-async function prepareESPCredentails(resource, sessionDuration){
+async function prepareESPCredentails(resource, sessionDuration) {
     let params = {
         region: config.region,
         durationInSeconds: sessionDuration
@@ -56,18 +56,18 @@ async function prepareESPCredentails(resource, sessionDuration){
     if (resource === AWSRESOURCE.S3) {
         params.iamRoleArn = config.s3RoleArn;
         params.externalId = config.s3ExternalId
-    } else if (resource === AWSRESOURCE.SQS){
+    } else if (resource === AWSRESOURCE.SQS) {
         params.iamRoleArn = config.sqsRoleArn;
         params.externalId = config.sqsExternalId
     }
-    
+
     let url = config.espTokenAuthorizeUrl;
-    const data = qs.stringify({token: config.espToken});
-    let configs = {headers: {"Content-Type": "application/x-www-form-urlencoded"}}
+    const data = qs.stringify({ token: config.espToken });
+    let configs = { headers: { "Content-Type": "application/x-www-form-urlencoded" } }
     const espToken = await axios.post(url, data, configs);
-    
+
     url = config.espAwsAuthorizeUrl
-    configs = {params: params, headers: {'Authorization': espToken.data.access_token} }
+    configs = { params: params, headers: { 'Authorization': espToken.data.access_token } }
     const res = await axios.get(url, configs);
 
     const Credentials = {
@@ -75,8 +75,8 @@ async function prepareESPCredentails(resource, sessionDuration){
         SecretAccessKey: res.data.secretAccessKey,
         SessionToken: res.data.sessionToken
     }
-    return {Credentials}
-    
+    return { Credentials }
+
 }
 
 module.exports = {
