@@ -7,79 +7,96 @@
 ***/
 
 
-const {ROLES, SPECAILCHARTOSTRING} = require('../config/constant');
+const { ROLES, SPECAILCHARTOSTRING } = require('../config/constant');
 const mongoDb = require('../db/mongo.db');
 const { ProjectModel, DataSetModel, UserModel } = require('../db/db-connect');
+const MESSAGE = require('../config/code_msg');
 
 function isASCII(str) {
     return /^[\x00-\xFF\u2013-\u2122]*$/.test(str);
 }
 
-function isNumeric(input){
+function isNumeric(input) {
     return /^-?[0-9]+.?[0-9]*/.test(input);
 }
 
-async function checkProjectByconditions(conditions, checkExsit){
+async function checkProjectByconditions(conditions, checkExsit) {
     const pro = await mongoDb.findByConditions(ProjectModel, conditions);
     if (checkExsit) {
         if (!pro[0]) {
-            throw {CODE: 4001, MSG: "NO PROJECT FOUND"};
+            throw MESSAGE.VALIDATATION_PJ_EMPTY;
         }
-    }else{
+    } else {
         if (pro[0]) {
-            throw {CODE: 4002, MSG: "PROJECT EXIST"};
+            throw MESSAGE.VALIDATATION_PJ_EXIST;
         }
     }
     return pro;
-    
+
 }
 
-async function checkAppendTicketsHeaders(appendHeaders, originalHeaders){
-    
-    const errRes = {CODE: 4005, MSG: "ERROR INPUT TICKET'S HEADERS"};
+async function checkAppendTicketsHeaders(appendHeaders, originalHeaders) {
+
     if (appendHeaders.length < originalHeaders.length) {
-        throw errRes;
+        throw MESSAGE.VALIDATION_TK_HEADEARS;
     }
-    originalHeaders.forEach(header =>{
+    originalHeaders.forEach(header => {
         if (!appendHeaders.includes(header)) {
-            throw errRes;
+            throw MESSAGE.VALIDATION_TK_HEADEARS;
         }
     });
 }
 
-async function checkUserRole(uid, checkRole){
+async function checkUser(uid, checkExsit) {
 
-   const user = await mongoDb.findById(UserModel, uid);
-   if (checkRole != user.role) {
-       throw {CODE: 4001, MSG: "ACCESS DENIED"}
-   }
+    const user = await mongoDb.findById(UserModel, uid);
+    if (checkExsit && !user) {
+        throw MESSAGE.VALIDATION_ACCESS;
+    }
+    return user;
 }
 
-async function checkAnnotator(uid){
+async function checkUserRole(uid, checkRole) {
+
+    const user = await mongoDb.findById(UserModel, uid);
+    if (checkRole != user.role) {
+        throw MESSAGE.VALIDATION_ACCESS;
+    }
+}
+
+async function checkAnnotator(uid) {
 
     const user = await mongoDb.findById(UserModel, uid);
     if (ROLES.ANNOTATOR == user.role) {
-        throw {CODE: 4001, MSG: "ACCESS DENIED"}
+        throw MESSAGE.VALIDATION_ACCESS;
     }
 }
 
-async function checkDataSet(conditions, checkExsit){
+async function checkAdmin(uid) {
+
+    const user = await mongoDb.findById(UserModel, uid);
+    if (ROLES.ADMIN != user.role) {
+        throw MESSAGE.VALIDATION_PERMITION;
+    }
+}
+
+async function checkDataSet(conditions, checkExsit) {
     const ds = await mongoDb.findByConditions(DataSetModel, conditions);
-     if (checkExsit) {
+    if (checkExsit) {
         if (!ds[0]) {
-            throw {CODE: 4001, MSG: "NO DATASET FOUND"};
+            throw MESSAGE.VALIDATION_DS_EMPTY;
         }
-    }else{
+    } else {
         if (ds[0]) {
-            throw {CODE: 4002, MSG: "DATASET EXIST"};
+            throw MESSAGE.VALIDATION_DS_EXIST;
         }
     }
     return ds;
- }
+}
 
 
 async function checkRequired(parameters) {
-    
+
     const paramType = typeof parameters;
     if (!parameters) {
         return false;
@@ -88,13 +105,13 @@ async function checkRequired(parameters) {
         if (parameters) {
             return true;
         }
-    }else if (paramType == SPECAILCHARTOSTRING.OBJECT) {
+    } else if (paramType == SPECAILCHARTOSTRING.OBJECT) {
         if (Object.keys(parameters).length) {
             return true;
         }
-    }else if (paramType == SPECAILCHARTOSTRING.STRING) {
+    } else if (paramType == SPECAILCHARTOSTRING.STRING) {
         const p = parameters.trim();
-        if (p && p != SPECAILCHARTOSTRING.ZERO && p != SPECAILCHARTOSTRING.FALSE  && p != SPECAILCHARTOSTRING.UNDEFINED && p != SPECAILCHARTOSTRING.NULL && p != SPECAILCHARTOSTRING.NAN ) {
+        if (p && p != SPECAILCHARTOSTRING.ZERO && p != SPECAILCHARTOSTRING.FALSE && p != SPECAILCHARTOSTRING.UNDEFINED && p != SPECAILCHARTOSTRING.NULL && p != SPECAILCHARTOSTRING.NAN) {
             return true;
         }
     }
@@ -103,22 +120,22 @@ async function checkRequired(parameters) {
 
 async function validateRequired(parameters) {
     if (!await checkRequired(parameters)) {
-     throw {4003: "input field invalid"}   
+        throw MESSAGE.VALIDATATION_INPUT;
     }
 }
 
 async function checkDataSetInUse(dataSetName, throwError) {
-    const conditions = { selectedDataset: dataSetName};
+    const conditions = { selectedDataset: dataSetName };
     const pro = await mongoDb.findByConditions(ProjectModel, conditions, "projectName");
     console.log(pro)
     if (pro[0] && throwError) {
-        const pnames = await pro.reduce((pnameString, curr) => pnameString?`${pnameString},${curr.projectName}`: curr.projectName, "");
-        throw {CODE: 4001, MSG: `DATA-SET USING BY: [${pnames}], PLEASE REMOVE THE PROJECTS FIRST.`};
+        const pnames = await pro.reduce((pnameString, curr) => pnameString ? `${pnameString},${curr.projectName}` : curr.projectName, "");
+        throw { CODE: MESSAGE.VALIDATION_DS_USING.CODE, MSG: MESSAGE.VALIDATION_DS_USING.MSG.replace('${pnames}', pnames) };
     }
     return pro;
 }
 
-async function checkJsonFormat(data){
+async function checkJsonFormat(data) {
     if (!data) {
         return false;
     }
@@ -132,11 +149,21 @@ async function checkJsonFormat(data){
     }
 }
 
+
+async function validateBool(data) {
+    if (data == true || data == 'true') {
+        return true;
+    }
+    return false;
+}
+
+
 module.exports = {
     isASCII,
     isNumeric,
     checkProjectByconditions,
     checkAppendTicketsHeaders,
+    checkUser,
     checkUserRole,
     checkAnnotator,
     checkDataSet,
@@ -144,5 +171,7 @@ module.exports = {
     validateRequired,
     checkDataSetInUse,
     checkJsonFormat,
-
+    validateBool,
+    checkAdmin,
+    
 };
